Fixes an issue where elements with z-index bleed through model
I came across an issue where I had some elements on the page that had
a z-index adn these were bleeding through the modal. To overcome this,
I added a z-index to the open when its in open state and increased the
index to a stupidly high number so its less likely other elements on
the page may ever bleed through again.