[MC,AK] Add a SN branded segmented controls comp
Initial commit

This is taken of a component designed by the Pegasus Team for the
Research Impacts application. It has been, in the end, trade with
another component (underline-nav), but it seemed to have raised interest
lately in the FED discipline and therefore maybe worth while getting
into Elements Design System.

Document is still Work In Progress!

Co-authored-by: Morgan Cugerone <morgan.cugerone@springernature.com>
Co-authored-by: Andreas Koscinski <andreas.koscinski@springernature.com>